Added super calculator with tests

+++ src/Employee.kt
@@ -1,3 +1,5 @@
-data class Employee(val firstName: String, val lastName: String, val salary: Int) {
+import java.math.BigDecimal
+
+data class Employee(val firstName: String, val lastName: String, val salary: BigDecimal) {
     fun combineFirstAndLastName() = "$firstName $lastName"
 }
