+++ src/PaymentInput.kt
@@ -1,3 +1,5 @@
-data class PaymentInput(val superRate: Int, val paymentStart: String, val paymentEnd: String) {
+import java.math.BigDecimal
+
+data class PaymentInput(val superRate: BigDecimal, val paymentStart: String, val paymentEnd: String) {
     fun combinePaymentStartAndEnd() = "$paymentStart - $paymentEnd"
 }
