+++ src/Payslip.kt
@@ -1,3 +1,5 @@
-data class Payslip(val employeeName: String, val payPeriod: String, val grossIncome: Int, val incomeTax: Int,
-              val netIncome: Int, val superAmt: Int) {
+import java.math.BigDecimal
+
+data class Payslip(val employeeName: String, val payPeriod: String, val grossIncome: BigDecimal, val incomeTax: BigDecimal,
+                   val netIncome: BigDecimal, val superAmt: BigDecimal) {
 }
